requirements.txt
asgiref==3.11.1
Django==4.2.7
django-filter==23.5
djangorestframework==3.14.0
psycopg2==2.9.11
python-dotenv==1.0.0
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2025.3
